% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tv_sf_data.R
\name{generate_tv_sf_data}
\alias{generate_tv_sf_data}
\title{generate_tv_sf_data}
\usage{
generate_tv_sf_data(p = 20, n = 2, N = 10, eta = 1.1, alpha = 0)
}
\arguments{
\item{p}{the number of variables. Default \code{20}.}

\item{n}{number of replicates. Default \code{2}.}

\item{N}{the number of time-points. Default \code{10}.}

\item{eta}{a constant controlling the positive definiteness of the matrix. Default \code{1.1}.}

\item{alpha}{parameter to control the sparsity of the network. Default \code{0}.}
}
\value{
A list of precision matrices, covariance matrices, and time-series data.
}
\description{
Simulate scale-free time-series data.
}
\examples{

set.seed(1)

n <- 5
p <- 50
N <- 100

Data <- generate_tv_sf_data(p = p, n = n, N = N)

}
