% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_glasso.R
\name{sf_glasso}
\alias{sf_glasso}
\title{sf_glasso}
\usage{
sf_glasso(S = NULL, alpha = 0.1, K = 2)
}
\arguments{
\item{S}{list of (smoothed) covariance matrices.}

\item{alpha}{the adaptive tuning parameter. Default \code{0.1}.}

\item{K}{the number of iterations (reweighting steps). Default \code{2}.}
}
\value{
A scale-free glasso estimate.
}
\description{
Computes scale-free glasso estimate
}
\examples{

library(glassoFast)
library(igraph)
library(MASS)

n <- 5
p <- 50
N <- 100

Data <- generate_tv_sf_data(p = p, n = n, N = N)

Y <- Data$X

pos <- 1:N
 
S_t <- kernel_cov_rep(X = Y, N = N, pos = pos)$S_t

Theta_est <- sf_glasso(S = S_t[[1]], K = 2)

}
