% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_cov_rep.R
\name{kernel_cov_rep}
\alias{kernel_cov_rep}
\title{kernel_cov_rep}
\usage{
kernel_cov_rep(X = NULL, N = NULL, pos = NULL, h = NULL, kernel = "gaussian")
}
\arguments{
\item{X}{the time-series data.}

\item{pos}{time points where network estimates are computed.}

\item{h}{the bandwidth. Default \code{5.848/N^(1/3)}.}

\item{kernel}{the kernel used. Default \code{gaussian}.}

\item{use_cor}{compute the correlation matrix. Default \code{FALSE}.}
}
\value{
List of smoothed covariance or correlation matrices.
}
\description{
Computes the smoothed covariance matrix estimates for biological replicates.
}
\examples{

library(huge)
library(igraph)
library(MASS)

n <- 5
p <- 100
N <- 100

Data <- generate_tv_sf_data(p = p, n = n, N = N)

Y <- Data$X

pos <- 1:N

S_t <- kernel_cov_rep(X = Y, N = N, pos = pos)

}
